+++ packages/faceted-search/stories/facetedSearch.story.js
@@ -21,6 +21,52 @@ while (i < 50) {
 	i += 1;
 }
 
+const badgesFaceted = {
+	badges: [
+		{
+			properties: {
+				attribute: 'connection.type',
+				initialOperatorOpened: false,
+				initialValueOpened: false,
+				label: 'Connection Type',
+				operator: {
+					label: 'In',
+					name: 'in',
+				},
+				operators: [
+					{
+						label: 'In',
+						name: 'in',
+					},
+				],
+				type: 'checkbox',
+				value: [
+					{
+						id: 'amazon_s3',
+						label: 'Amazon S3',
+						checked: true,
+					},
+				],
+			},
+			metadata: {
+				badgePerFacet: '1',
+				entitiesPerBadge: 'N',
+				values: [
+					{ id: 'amazon_s3', label: 'Amazon S3' },
+					{ id: 'hdfs', label: 'HDFS' },
+					{ id: 'kafka', label: 'Kafka' },
+					{ id: 'localcon', label: 'Local connection' },
+					{ id: 'salesforce', label: 'Salesforce' },
+					{ id: 'aws_kinesis', label: 'AWS kinesis' },
+				],
+				operators: ['in'],
+				badgeId: 'connection.type-9f0e5bc7-c687-4198-9635-d0fc7724dfd1',
+				isInCreation: false,
+			},
+		},
+	],
+};
+
 const paddingLeft = { paddingLeft: '10px' };
 
 storiesOf('FacetedSearch', module)
@@ -57,6 +103,25 @@ storiesOf('FacetedSearch', module)
 					(currentFacetedMode === FacetedSearch.constants.FACETED_MODE.BASIC && (
 						<FacetedSearch.BasicSearch
 							badgesDefinitions={badgesDefinitions}
+							onSubmit={action('onSubmit')}
+						/>
+					))
+				}
+			</FacetedSearch.Faceted>
+		</div>
+	))
+	.add('initialized', () => (
+		<div>
+			<IconsProvider />
+			<FacetedSearch.Faceted id="my-faceted-search">
+				{currentFacetedMode =>
+					(currentFacetedMode === FacetedSearch.constants.FACETED_MODE.ADVANCED && (
+						<FacetedSearch.AdvancedSearch onSubmit={action('onSubmit')} />
+					)) ||
+					(currentFacetedMode === FacetedSearch.constants.FACETED_MODE.BASIC && (
+						<FacetedSearch.BasicSearch
+							badgesDefinitions={badgesDefinitions}
+							badgesFaceted={badgesFaceted}
 							onSubmit={action('onSubmit')}
 						/>
 					))